test(desktop): keep proxy coverage below debt ceiling
Problem:
- the IPv6 proxy-host regression pushed settings_app_test.go above its repolint budget in CI

Root cause:
- proxy-specific tests were left in an already oversized legacy test file

Fix:
- move the proxy persistence and direct-host tests into network_proxy_app_test.go
- lower the repolint totals and restore settings_app_test.go to its main-v2 baseline

Verification:
- go test . -run TestNetworkProxySpecForRootMatchesEffectiveProjectConfig\|TestWithProbeDirectHostMirrorsProviderNoProxy\|TestSaveProviderPersistsNoProxy -count=1 (desktop)
- go run ./tools/repolint
- git diff --check
